combines two KG examples into one
