+++ examples/knowledge_graph/requirements.txt
@@ -1,3 +1,5 @@
 indexify
 neo4j
+spacy
+sentence-transformers
 google-generativeai
