--- examples/knowledge_graph_generation/requirements.txt
@@ -1,4 +1,0 @@
-indexify
-neo4j
-spacy
-sentence-transformers
